+++ crates/icu/src/ic/mod.rs
@@ -1,16 +1,30 @@
 ///
 /// IMPORT IC CRATES
 ///
+pub use ic_cdk::*;
 
 pub mod api {
     pub use ic_cdk::api::*;
 }
+
+pub mod candid {
+    pub use candid::*;
+}
+
 pub mod mgmt {
     pub use ic_cdk::management_canister::*;
 }
-pub use ic_cdk::*;
-pub mod icrc {
+
+pub mod ledger_types {
+    pub use ic_ledger_types::*;
+}
+
+pub mod icrc_ledger_types {
     pub use icrc_ledger_types::*;
+}
+
+pub mod principal {
+    pub use ic_principal::*;
 }
 
 pub mod timers {