+++ crates/icu/src/lib.rs
@@ -26,7 +26,10 @@ use thiserror::Error as ThisError;
 pub mod prelude {
     pub use crate::{
         Error as IcuError, Log, auth_require_all, auth_require_any,
-        ic::{api::msg_caller, export_candid, init, query, update},
+        ic::{
+            api::msg_caller, candid::CandidType, export_candid, init, ledger_types::Subaccount,
+            principal::Principal, query, update,
+        },
         icu_register_memory, icu_start, icu_start_root, log, perf,
     };
 }