+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable, and occasionally less notable changes to this project will be docum
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.1.5]
+- not using candid Principal any more, switching to ic_principal and ic_ledger_types
+
 ## [0.1.4]
 - refactored into two crates, just so I have a test crate to play with
 - updated icu_start! so it takes another optional argument to pass to the init function